monitoring/security: Hide delete downtime action in the downtimes overview if user lacks the respective permission

+++ modules/monitoring/application/controllers/ListController.php
@@ -306,7 +306,10 @@ class Monitoring_ListController extends Controller
         ));
 
         $this->view->downtimes = $query->paginate();
-        $this->view->delDowntimeForm = new DeleteDowntimeCommandForm();
+
+        if ($this->Auth()->hasPermission('monitoring/command/downtime/delete')) {
+            $this->view->delDowntimeForm = new DeleteDowntimeCommandForm();
+        }
     }
 
     /**
